Introducing Maître D` as a gate keeper (#4)
* Maître D is meant to reduce the cognitive complexity of reservations controller.
 Furthermore it enables more reservation on the same day for possible more tables.

The table allocation happens will requesting existing reserations. Therefore only 
reservations are stored and kept. The new reservation is checked against existing
reservations. If all overlapping reservations and the new reservation are able to find
an existing and usable table with the right size. Then the reservation is allowed and
will be stored at the reservation repository as well.
---------

Co-authored-by: Torsten Compart <torsten.compart@gmail.com>